MessageLayerを別クラスに分離
